requirements.txt
google-generativeai>=0.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
dnspython>=2.4.0
flask>=2.3.0
werkzeug>=2.3.0

# Matchmaking Users dependencies
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.3.2
numpy==1.24.3
